- code fixes and moves compose theme to compose package
